Add files via upload
This is my second project that i made from my Udemy WebDev course : TinDog i.e. tinder for dogs 😆.
Its basically what it says this is my second project and i think i am much more satisfied with it than the previous one. Still some things that are off are that the images on the carousel are of different dimensions and hence the carousel changes its size when the forward buttons are used which spoils the overall look
Other than this, ofcourse the site has no functionality, just basic html css so its no fun 
Lets hope for a better next pro 🤞